Add command to reporting potentially approvable patches
Add a 'gerrymander todo-approvable' command which allows
a core team reviewer to identify patches which they have
not commented on before, which are candidates for being
approved to merge.

A patch is potentially approvable if it has at least one
+2 vote already and does not have a -2, or a workflow
vote of -1 or +1. To be very strict it is possible to
also ignore any patches with a vote of -1, though it is
valid to ignore -1 code review votes, so make this filter
optional via a '--strict' flag.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>